Re-organise yaml checks code
The core.checks module is very monolithic so this
pulls code related to yaml checks into a new core.ycheck
module that contains a submodule per type of check i.e.
bugs, events etc.
