+++ core/checks.py
@@ -1,709 +1,17 @@
 import os
 
-import importlib
-import operator
 import re
 import subprocess
-import yaml
 
 from core import constants
-from core.issues import issue_utils
 from core.cli_helpers import CLIHelper
-from core.log import log
-from core.utils import mktemp_dump, sorted_dict
-from core.known_bugs_utils import (
-    add_known_bug,
-    BugSearchDef,
-)
-from core.searchtools import (
-    FileSearcher,
-    SearchDef,
-    SequenceSearchDef,
-)
-from core.ystruct import (
-    YAMLDefOverrideBase,
-    YAMLDefSection,
-)
+from core.utils import sorted_dict
 
 SVC_EXPR_TEMPLATES = {
     "absolute": r".+\S+bin/({})(?:\s+.+|$)",
     "snap": r".+\S+\d+/({})(?:\s+.+|$)",
     "relative": r".+\s({})(?:\s+.+|$)",
     }
-
-
-class YAMLDefExpr(YAMLDefOverrideBase):
-    KEYS = ['start', 'body', 'end', 'expr', 'hint']
-
-    @property
-    def expr(self):
-        """
-        Subkey e.g for start.expr, body.expr. Expression defs that are just
-        a string or use subkey 'expr' will rely on __getattr__.
-        """
-        return self.content.get('expr', self.content)
-
-    @property
-    def hint(self):
-        return self.content.get('hint', None)
-
-    def __getattr__(self, name):
-        """
-        An expression can be provided in multiple ways. The simplest form is
-        where the value of the event is a pattern string hence why we override
-        this to just return the content if it is not a dict.
-        """
-        if type(self.content) == dict:
-            return super().__getattr__(name)
-        else:
-            return self.content
-
-
-class YAMLDefMessage(YAMLDefOverrideBase):
-    KEYS = ['message', 'message-format-result-groups']
-
-    @property
-    def format_groups(self):
-        return self.content
-
-    def __str__(self):
-        return self.content
-
-
-class YAMLDefSettings(YAMLDefOverrideBase):
-    KEYS = ['settings']
-
-    def __iter__(self):
-        for key, val in self.content.items():
-            yield key, val
-
-
-class YAMLDefReleases(YAMLDefOverrideBase):
-    KEYS = ['releases']
-
-    def __iter__(self):
-        for key, val in self.content.items():
-            yield key, val
-
-
-class YAMLDefResultsPassthrough(YAMLDefOverrideBase):
-    KEYS = ['passthrough-results']
-
-    def __bool__(self):
-        """ If returns True, the master results list is passed to callbacks
-        so that they may fetch results in their own way. This is required e.g.
-        when processing results with analytics.LogEventStats.
-        """
-        return bool(self.content)
-
-
-class YAMLDefInput(YAMLDefOverrideBase):
-    KEYS = ['input']
-    TYPE_COMMAND = 'command'
-    TYPE_FILESYSTEM = 'filesystem'
-
-    # NOTE: this must be set by the handler object that is using the overrides.
-    #       we need a better way to do this but we can't use __init__ because
-    #       this class must be provided uninstantiated.
-    EVENT_CHECK_OBJ = None
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._path = None
-
-    @property
-    def meta(self):
-        defaults = {'allow-all-logs': True,
-                    'args': [],
-                    'kargs': {},
-                    'args-callback': None}
-        _meta = self.content.get('meta', defaults)
-        defaults.update(_meta)
-        return defaults
-
-    @property
-    def path(self):
-        if self.type == self.TYPE_FILESYSTEM:
-            path = os.path.join(constants.DATA_ROOT, self.value)
-            if constants.USE_ALL_LOGS and self.meta['allow-all-logs']:
-                path = "{}*".format(path)
-
-            return path
-        elif self.type == self.TYPE_COMMAND:
-            if self._path:
-                return self._path
-
-            args_callback = self.meta['args-callback']
-            if args_callback:
-                args, kwargs = getattr(self.EVENT_CHECK_OBJ, args_callback)()
-            else:
-                args = self.meta['args']
-                kwargs = self.meta['kwargs']
-
-            # get command output
-            out = getattr(CLIHelper(), self.value)(*args,
-                                                   **kwargs)
-            # store in temp file to make it searchable
-            # NOTE: we dont need to delete this at the the end since they are
-            # created in the plugun tmp dir which is wiped at the end of the
-            # plugin run.
-            self._path = mktemp_dump(''.join(out))
-            return self._path
-
-
-class YAMLDefRequires(YAMLDefOverrideBase):
-    KEYS = ['requires']
-
-    @property
-    def passes(self):
-        if self.type == 'apt':
-            pkg = self.value
-            result = APTPackageChecksBase(core_pkgs=[pkg]).is_installed(pkg)
-            log.debug('config_check requirement:apt %s (result=%s)', pkg,
-                      result)
-            return result
-        elif self.type == 'property':
-            mod = self.source.rpartition('.')[0]
-            property = self.source.rpartition('.')[2]
-            class_name = mod.rpartition('.')[2]
-            mod = mod.rpartition('.')[0]
-            cls = getattr(importlib.import_module(mod), class_name)
-            result = getattr(cls(), property) is self.value
-            log.debug('config_check requirement:property %s is %s (result=%s)',
-                      self.source, self.value, result)
-            return result
-        else:
-            self.debug("unsupported yaml requires type=%s", self.type)
-
-        return False
-
-
-class YAMLDefIssueType(YAMLDefOverrideBase):
-    KEYS = ['raises']
-
-    @property
-    def issue(self):
-        mod = self.content.rpartition('.')[0]
-        class_name = self.content.rpartition('.')[2]
-        return getattr(importlib.import_module(mod), class_name)
-
-
-class YAMLDefConfig(YAMLDefOverrideBase):
-    KEYS = ['config']
-
-    def actual(self, key, section=None):
-        mod = self.handler.rpartition('.')[0]
-        cls = self.handler.rpartition('.')[2]
-        obj = getattr(importlib.import_module(mod), cls)
-        if hasattr(self, 'path'):
-            self.cfg = obj(self.path)
-        else:
-            self.cfg = obj()
-
-        if section:
-            actual = self.cfg.get(key, section=section)
-        else:
-            actual = self.cfg.get(key)
-
-        return actual
-
-    def check(self, actual, value, op, allow_unset=False):
-        if value is not None and actual is None:
-            if allow_unset:
-                return True
-            else:
-                return False
-
-        # Apply the type from the yaml to that of the config
-        if value is not None and type(value) != type(actual):
-            actual = type(value)(actual)
-
-        if getattr(operator, op)(actual, value):
-            return True
-
-        return False
-
-
-class CallbackHelper(object):
-
-    def __init__(self):
-        self.callbacks = {}
-
-    def callback(self, f):
-        def callback_inner(*args, **kwargs):
-            return f(*args, **kwargs)
-
-        self.callbacks[f.__name__] = callback_inner
-        # we don't need to return but we leave it so that we can unit test
-        # these methods.
-        return callback_inner
-
-
-class ChecksBase(object):
-
-    def __init__(self, *args, yaml_defs_group=None, searchobj=None, **kwargs):
-        """
-        @param _yaml_defs_group: optional key used to identify our yaml
-                                 definitions if indeed we have any. This is
-                                 given meaning by the implementing class.
-        @param searchobj: optional FileSearcher object used for searches. If
-                          multiple implementations of this class are used in
-                          the same part it is recommended to provide a search
-                          object that is shared across them to provide
-                          concurrent execution.
-
-        """
-        super().__init__(*args, **kwargs)
-        if searchobj:
-            self.searchobj = searchobj
-        else:
-            self.searchobj = FileSearcher()
-
-        self._yaml_defs_group = yaml_defs_group
-
-    def register_search_terms(self):
-        raise NotImplementedError
-
-    def process_results(self, results):
-        raise NotImplementedError
-
-    def run_checks(self):
-        self.register_search_terms()
-        return self.process_results(self.searchobj.search())
-
-
-class PackageReleaseCheckObj(object):
-    def __init__(self, package_name):
-        self.package_name = package_name
-        self.bugs = {}
-
-    def add_bug_check(self, id, release, minbroken, minfixed, message):
-        bug = {'id': id, 'minbroken': minbroken, 'minfixed': minfixed,
-               'message': message}
-        if release in self.bugs:
-            self.bugs[release].append(bug)
-        else:
-            self.bugs[release] = [bug]
-
-
-class PackageBugChecksBase(object):
-    """
-    This is used to check if the version of installed packages contain
-    known bugs and report them if found.
-    """
-    def __init__(self, release_name, pkg_info):
-        """
-        @param release_name: release name we are checking against - this must
-                             exist in the yaml defs.
-        @param pkg_info: dict of installed packages in the for
-                         {<name>: <version>}. This is typically obtained from
-                         an implementation of APTPackageChecksBase.
-        """
-        self._release_name = release_name
-        self._pkg_info = pkg_info
-        self._checks = []
-
-    def _load_definitions(self):
-        """
-        Load bug search definitions from yaml.
-
-        @return: list of BugSearchDef objects
-        """
-        path = os.path.join(constants.PLUGIN_YAML_DEFS,
-                            'package_bug_checks.yaml')
-        with open(path) as fd:
-            yaml_defs = yaml.safe_load(fd.read())
-
-        if not yaml_defs:
-            return
-
-        overrides = [YAMLDefInput, YAMLDefExpr, YAMLDefMessage,
-                     YAMLDefReleases]
-        # TODO: need a better way to provide this instance to the input
-        #       override.
-        YAMLDefInput.EVENT_CHECK_OBJ = self
-        plugin = yaml_defs.get(constants.PLUGIN_NAME, {})
-        group = YAMLDefSection(constants.PLUGIN_NAME, plugin,
-                               override_handlers=overrides)
-        log.debug("sections=%s, events=%s",
-                  len(group.branch_sections),
-                  len(group.leaf_sections))
-
-        for bug_check in group.leaf_sections:
-            pkg_name = bug_check.parent.name
-            p = PackageReleaseCheckObj(pkg_name)
-            bug = bug_check.name
-            message = str(bug_check.message)
-            for rel, info in dict(bug_check.releases).items():
-                p.add_bug_check(bug, rel, info['min-broken'],
-                                info['min-fixed'], message)
-
-            self._checks.append(p)
-
-    def __call__(self):
-        self._load_definitions()
-        for check in self._checks:
-            pkg = check.package_name
-            # if installed do check
-            if pkg not in self._pkg_info:
-                return
-
-            pkgver = self._pkg_info[pkg]
-            for release, bugs in check.bugs.items():
-                if release != self._release_name:
-                    continue
-
-                for bug in bugs:
-                    minbroken = bug['minbroken']
-                    minfixed = bug['minfixed']
-                    if (not pkgver < DPKGVersionCompare(minbroken) and
-                            pkgver < DPKGVersionCompare(minfixed)):
-                        log.debug("bug identified for package=%s release=%s "
-                                  "version=%s", pkg, release, pkgver)
-                        message_format_kwargs = {'package_name': pkg,
-                                                 'version_current': pkgver,
-                                                 'version_fixed': minfixed}
-
-                        message = bug.get('message')
-                        if not message:
-                            # generic message
-                            message = ("package {package_name} with version "
-                                       "{version_current} contains a known "
-                                       "bug and should be upgraded to >= "
-                                       "{version_fixed}")
-
-                        add_known_bug(bug['id'],
-                                      message.format(**message_format_kwargs))
-
-
-class BugChecksBase(ChecksBase):
-    """
-    Class used to identify bugs by searching for entries in files (typically
-    log files). What we search for and how we respond is defined in
-    defs/bugs.yaml so that we can grow our searches without touching code.
-
-    Searches are defined per plugin and run automatically i.e. there is no need
-    to implement this class.
-    """
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._bug_defs = None
-
-    def _load_bug_definitions(self):
-        """
-        Load bug search definitions from yaml.
-
-        @return: list of BugSearchDef objects
-        """
-        path = os.path.join(constants.PLUGIN_YAML_DEFS, "bugs.yaml")
-        with open(path) as fd:
-            yaml_defs = yaml.safe_load(fd.read())
-
-        if not yaml_defs:
-            return
-
-        plugin_bugs = yaml_defs.get(constants.PLUGIN_NAME, {})
-        overrides = [YAMLDefInput, YAMLDefExpr, YAMLDefMessage]
-        # TODO: need a better way to provide this instance to the input
-        #       override.
-        YAMLDefInput.EVENT_CHECK_OBJ = self
-        plugin = YAMLDefSection(constants.PLUGIN_NAME, plugin_bugs,
-                                override_handlers=overrides)
-        log.debug("loading plugin '%s' bugs - sections=%s, events=%s",
-                  plugin.name,
-                  len(plugin.branch_sections),
-                  len(plugin.leaf_sections))
-        bug_defs = []
-        for bug in plugin.leaf_sections:
-            id = bug.name
-            message_format = bug.message_format_result_groups
-            if message_format:
-                message_format = message_format.format_groups
-
-            pattern = bug.expr.value
-            # NOTE: pattern can be string or list of strings
-            bdef = {'def':
-                    BugSearchDef(
-                            pattern,
-                            bug_id=str(id),
-                            hint=bug.hint.value,
-                            message=str(bug.message),
-                            message_format_result_groups=message_format)}
-
-            datasource = bug.input.path
-            log.debug("bug=%s path=%s", id, datasource)
-            bdef["datasource"] = datasource
-            bug_defs.append(bdef)
-
-        self._bug_defs = bug_defs
-
-    @property
-    def bug_definitions(self):
-        """
-        @return: dict of SearchDef objects and datasource for all entries in
-        bugs.yaml under _yaml_defs_group.
-        """
-        if self._bug_defs is not None:
-            return self._bug_defs
-
-        self._load_bug_definitions()
-        return self._bug_defs
-
-    def register_search_terms(self):
-        for bugsearch in self.bug_definitions:
-            self.searchobj.add_search_term(bugsearch["def"],
-                                           bugsearch["datasource"])
-
-    def process_results(self, results):
-        for bugsearch in self.bug_definitions:
-            tag = bugsearch["def"].tag
-            _results = results.find_by_tag(tag)
-            if _results:
-                if bugsearch["def"].message_format_result_groups:
-                    message = bugsearch["def"].rendered_message(_results[0])
-                    add_known_bug(tag, message)
-                else:
-                    add_known_bug(tag, bugsearch["def"].message)
-
-    def __call__(self):
-        """ Must be callable since its run automatically. """
-        self.run_checks()
-
-
-class EventCheckResult(object):
-    """ This is passed to an event check callback when matches are found """
-
-    def __init__(self, defs_section, defs_event, search_results,
-                 sequence_def=None):
-        """
-        @param defs_section: section name from yaml
-        @param defs_event: event label/name from yaml
-        @param search_results: searchtools.SearchResultsCollection
-        @param sequence_def: if set the search results are from a
-                            searchtools.SequenceSearchDef and are therefore
-                            grouped as sections of results rather than a single
-                            set of results.
-        """
-        self.section = defs_section
-        self.name = defs_event
-        self.results = search_results
-        self.sequence_def = sequence_def
-
-
-class EventChecksBase(ChecksBase):
-
-    def __init__(self, *args, callback_helper=None,
-                 event_results_output_key=None, **kwargs):
-        """
-        @param callback_helper: optionally provide a callback helper. This is
-        used to "register" callbacks against events defined in the yaml so
-        that they are automatically called when corresponding events are
-        detected.
-        @param event_results_output_key: by default the plugin output will be
-                                         set to the return of process_results()
-                                         but that can be optionally set
-                                         with this key as root e.g. to avoid
-                                         clobbering other results.
-        """
-        super().__init__(*args, **kwargs)
-        self.callback_helper = callback_helper
-        self.event_results_output_key = event_results_output_key
-        self._event_defs = {}
-
-    def _load_event_definitions(self):
-        """
-        Load event search definitions from yaml.
-
-        An event is identified using between one and two expressions. If it
-        requires a start and end to be considered complete then these can be
-        specified for match otherwise we can match on a single line.
-        Note that multi-line events can be overlapping hence why we don't use a
-        SequenceSearchDef (we use core.analytics.LogEventStats).
-        """
-        path = os.path.join(constants.PLUGIN_YAML_DEFS, "events.yaml")
-        with open(path) as fd:
-            yaml_defs = yaml.safe_load(fd.read())
-
-        if not yaml_defs:
-            return
-
-        log.debug("loading event definitions for plugin=%s group=%s",
-                  constants.PLUGIN_NAME, self._yaml_defs_group)
-        plugin = yaml_defs.get(constants.PLUGIN_NAME, {})
-        group_name = self._yaml_defs_group
-
-        overrides = [YAMLDefInput, YAMLDefExpr, YAMLDefResultsPassthrough]
-        # TODO: need a better way to provide this instance to the input
-        #       override.
-        YAMLDefInput.EVENT_CHECK_OBJ = self
-        group = YAMLDefSection(group_name, plugin.get(group_name),
-                               override_handlers=overrides)
-        log.debug("sections=%s, events=%s",
-                  len(group.branch_sections),
-                  len(group.leaf_sections))
-
-        for event in group.leaf_sections:
-            results_passthrough = bool(event.passthrough_results)
-            log.debug("event: %s", event.name)
-            log.debug("input: %s:%s", event.input.type, event.input.path)
-            log.debug("passthrough: %s", results_passthrough)
-
-            # if this is a multiline event (has a start and end), append
-            # this to the tag so that it can be used with
-            # core.analytics.LogEventStats.
-            search_meta = {'searchdefs': [], 'datasource': None,
-                           'passthrough_results': results_passthrough}
-
-            if event.expr:
-                hint = None
-                if event.hint:
-                    hint = event.hint.value
-
-                search_meta['searchdefs'].append(
-                    SearchDef(event.expr.value, tag=event.name, hint=hint))
-            elif event.start:
-                if (event.body or
-                        (event.end and not results_passthrough)):
-                    log.debug("event '%s' search is a sequence", event.name)
-                    sd_start = SearchDef(event.start.expr)
-
-                    sd_end = None
-                    # explicit end is optional for sequence definition
-                    if event.end:
-                        sd_end = SearchDef(event.end.expr)
-
-                    sd_body = None
-                    if event.body:
-                        sd_body = SearchDef(event.body.expr)
-
-                    # NOTE: we don't use hints here
-                    sequence_def = SequenceSearchDef(start=sd_start,
-                                                     body=sd_body,
-                                                     end=sd_end,
-                                                     tag=event.name)
-                    search_meta['searchdefs'].append(sequence_def)
-                    search_meta['is_sequence'] = True
-                elif (results_passthrough and
-                      (event.start and event.end)):
-                    # start and end required for core.analytics.LogEventStats
-                    search_meta['searchdefs'].append(
-                        SearchDef(event.start.expr,
-                                  tag="{}-start".format(event.name),
-                                  hint=event.start.hint))
-                    search_meta['searchdefs'].append(
-                        SearchDef(event.end.expr,
-                                  tag="{}-end".format(event.name),
-                                  hint=event.end.hint))
-                else:
-                    log.debug("unexpected search definition passthrough=%s "
-                              "body provided=%s, end provided=%s",
-                              results_passthrough, event.body is not None,
-                              event.end is not None)
-            else:
-                log.debug("invalid search definition for event '%s' in "
-                          "section '%s'", event, event.parent.name)
-                continue
-
-            datasource = event.input.path
-            section_name = event.parent.name
-            if section_name not in self._event_defs:
-                self._event_defs[section_name] = {}
-
-            search_meta['datasource'] = datasource
-            self._event_defs[section_name][event.name] = search_meta
-
-    @property
-    def event_definitions(self):
-        """
-        @return: dict of SearchDef objects and datasource for all entries in
-        events.yaml under _yaml_defs_group.
-        """
-        if self._event_defs:
-            return self._event_defs
-
-        self._load_event_definitions()
-        return self._event_defs
-
-    def register_search_terms(self):
-        """
-        Register the search definitions for all events.
-
-        @param root_key: events.yaml root key
-        """
-        for defs in self.event_definitions.values():
-            for label in defs:
-                event = defs[label]
-                for sd in event["searchdefs"]:
-                    self.searchobj.add_search_term(sd, event["datasource"])
-
-    def process_results(self, results):
-        """
-        Provide a default way for results to be processed. This requires a
-        CallbackHelper to have been provided and callbacks registered. If that
-        is not the case the method must be re-implemented with another means
-        of processing results.
-
-        See defs/events.yaml for definitions.
-        """
-        if self.callback_helper is None or not self.callback_helper.callbacks:
-            # If there are no callbacks registered this method must be
-            # (re)implemented.
-            raise NotImplementedError
-
-        info = {}
-        for section_name, section in self.event_definitions.items():
-            for event, event_meta in section.items():
-                sequence_def = None
-                if event_meta.get('passthrough_results'):
-                    # this is for implementations that have their own means of
-                    # retreiving results.
-                    search_results = results
-                else:
-                    if event_meta.get('is_sequence'):
-                        sequence_def = event_meta['searchdefs'][0]
-                        search_results = results.find_sequence_sections(
-                            sequence_def)
-                        if search_results:
-                            search_results = search_results.values()
-                    else:
-                        search_results = results.find_by_tag(event)
-
-                if not search_results:
-                    continue
-
-                # We want this to throw an exception if the callback is not
-                # defined.
-                callback_name = event.replace('-', '_')
-                callback = self.callback_helper.callbacks[callback_name]
-                log.debug("executing event callback '%s'", callback_name)
-                event_results_obj = EventCheckResult(section_name, event,
-                                                     search_results,
-                                                     sequence_def=sequence_def)
-                ret = callback(self, event_results_obj)
-                if not ret:
-                    continue
-
-                # if the return is a tuple it is assumed to be of the form
-                # (<output value>, <output key>) where <output key> is used to
-                # override the output key for the result which defaults to the
-                # event name.
-                if type(ret) == tuple:
-                    out_key = ret[1]
-                    ret = ret[0]
-                else:
-                    out_key = event
-
-                # Don't clobber results, instead allow them to aggregate.
-                if out_key in info:
-                    info[out_key].update(ret)
-                else:
-                    info[out_key] = ret
-
-        if info:
-            if self.event_results_output_key:
-                info = {self.event_results_output_key: info}
-
-            return info
 
 
 class ConfigBase(object):
@@ -838,99 +146,6 @@ class SectionalConfigBase(ConfigBase):
                     val = constants.bool_str(ret.group(2))
                     self._sections[current_section][key] = val
                     self._flattened_config[key] = val
-
-
-class ConfigChecksBase(object):
-    """
-    This class is used to peform checks on file based config. The input is
-    defined in defs/config_checks.yaml.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._check_defs = {}
-
-    def _load_definitions(self):
-        path = os.path.join(constants.PLUGIN_YAML_DEFS, "config_checks.yaml")
-        with open(path) as fd:
-            yaml_defs = yaml.safe_load(fd.read())
-
-        if not yaml_defs:
-            return
-
-        log.debug("loading config check definitions for plugin '%s'",
-                  constants.PLUGIN_NAME)
-        overrides = [YAMLDefInput, YAMLDefExpr, YAMLDefMessage,
-                     YAMLDefRequires, YAMLDefConfig, YAMLDefSettings,
-                     YAMLDefIssueType]
-        # TODO: need a better way to provide this instance to the input
-        #       override.
-        YAMLDefInput.EVENT_CHECK_OBJ = self
-        plugin = yaml_defs.get(constants.PLUGIN_NAME, {})
-        group = YAMLDefSection(constants.PLUGIN_NAME, plugin,
-                               override_handlers=overrides)
-        log.debug("sections=%s, events=%s",
-                  len(group.branch_sections),
-                  len(group.leaf_sections))
-
-        for cfg_check in group.leaf_sections:
-            # This is only available if there is a section above us
-            if cfg_check.parent:
-                group_name = cfg_check.parent.name
-            else:
-                group_name = None
-
-            self._check_defs[cfg_check.name] = {
-                'group': group_name,
-                'message': cfg_check.message,
-                'config': cfg_check.config,
-                'requires': cfg_check.requires,
-                'settings': dict(cfg_check.settings),
-                'raises': cfg_check.raises}
-
-    def run_checks(self):
-        for name, cfg_check in self._check_defs.items():
-            # this is optional
-            requires = cfg_check['requires']
-            if requires and not cfg_check['requires'].passes:
-                continue
-
-            log.debug("config check section=%s", name)
-            message = None
-            if cfg_check['message']:
-                message = str(cfg_check['message'])
-
-            for cfg_key, settings in cfg_check['settings'].items():
-                op = settings['operator']
-                value = settings['value']
-                section = settings.get('section')
-                allow_unset = bool(settings.get('allow-unset', False))
-                raise_issue = False
-                actual = cfg_check['config'].actual(cfg_key, section=section)
-                log.debug("checking config %s %s %s (actual=%s)", cfg_key, op,
-                          value, actual)
-                if not cfg_check['config'].check(actual, value, op,
-                                                 allow_unset=allow_unset):
-                    raise_issue = True
-
-                if raise_issue:
-                    if message:
-                        _message = message
-                    else:
-                        _message = ("{} config {} expected to be {} {} "
-                                    "but actual={}".format(cfg_check['group'],
-                                                           cfg_key, op, value,
-                                                           actual))
-
-                    issue = cfg_check['raises'].issue(_message)
-                    issue_utils.add_issue(issue)
-                    # move on to next set of checks
-                    break
-
-    def __call__(self):
-        """ Must be callable since it's run automatically. """
-        self._load_definitions()
-        self.run_checks()
 
 
 class ServiceChecksBase(object):
