+++ defs/events.yaml
@@ -9,7 +9,7 @@
 #       EVENT:
 #         SETTINGS
 #
-# Supported SETTINGS (for more details see common.checks.EventChecksBase):
+# Supported SETTINGS (for more details see core.ycheck.YEventCheckerBase):
 #
 # Two types of searches are available here; single or multi line. A multi-line
 # search can be used in two ways; the first use is simply to have a "start" and
