+++ defs/package_bug_checks.yaml
@@ -24,6 +24,9 @@
 #   - version_fixed
 #
 openstack:
+  context:
+    release: core.plugins.openstack.OpenstackBase.release_name
+    apt-all: core.plugins.openstack.OpenstackBase.apt_packages_all
   neutron-common:
     1927868:
       message: >-
